updates and bulk commit because i cba to individually commit for each file lol

+++ LavalinkServer/src/main/java/lavalink/server/io/WebSocketHandlers.kt
@@ -93,16 +93,16 @@ class WebSocketHandlers(private val contextMap: Map<String, SocketContext>) {
 
   fun volume(context: SocketContext, json: JSONObject) {
     val player = context.getPlayer(json.getString("guildId"))
-    player.setVolume(json.getInt("volume"))
+    player.setVolume(json.getFloat("volume"))
   }
 
   fun equalizer(context: SocketContext, json: JSONObject) {
     if (!loggedEqualizerDeprecationWarning) log.warn("The 'equalizer' op has been deprecated in favour of the " +
       "'filters' op. Please switch to use that one, as this op will get removed in v4.")
+
     loggedEqualizerDeprecationWarning = true
 
     val player = context.getPlayer(json.getString("guildId"))
-
     val list = mutableListOf<Band>()
     json.getJSONArray("bands").forEach { b ->
       val band = b as JSONObject
